==> 00-intro.qmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Introduction {.unnumbered}

## Démarche

Il s'agit d'un support de formation conçu pour initier à l'économétrie spatiale chercheurs, doctorants et étudiants s'orientant vers une thèse. Ce document a été élaboré avec les priorités suivantes :

-   Compréhension des enjeux théoriques et méthodologiques soulevés par les évaluations d'impact ;
-   Identification des spécificités et opportunités liées à l'évaluation d'interventions spatialisées ;
-   Obtention de données spatiale permettant de démontrer et de quantifier l'impact ;
-   Compréhension des différentes approches pour constituer un contrefactuel : randomisation, appariement, comparaison avant-après, différence de différences ;
-   Familiarisation avec l'outil de traitement statistique R;
-   Développement d'un regard critique sur la pertinence et la fiabilité des données mobilisées.

D'autres aspects ont volontairement été placés au second plan, en particulier les subtilités associées aux modèles économétriques. Les méthodes économétriques mobilisées ici sont volontairement simples : pour être publiable dans une revue scientifique, un travail d'évaluation doit mobiliser des spécifications plus sophistiquées.

## Cadrage général pour les méthodes d'évaluation

```{=html}
<iframe src="/presentations/TVM_Intro_generale.pdf" width="100%" height="400px"></iframe>
```

[Cliquer ici pour télécharger la présentation](/presentations/TVM_Intro_generale.pdf)


## Revue des articles
