==> 03_assignation_aleatoire.qmd <==
---
title: "Données spatiales"
author: "Florent Bédécarrats"
format: html
editor: visual
execute: 
  eval: false
editor_options: 
  chunk_output_type: console
---

# Méthode randomisée

```{=html}
<iframe src="presentations/TVM_RCT.pdf" width="100%" height="400px"></iframe>
```
[Cliquer ici pour télécharger la présentation](presentations/TVM_RCT.pdf).

> **ATTENTION :** Il va de soi que les AP malgaches n'ont à aucun moment été assignées aléatoirement. Lors de cette séquence, on fait "comme si", pour montrer la manière dont les données sont analysées quand il y a eu assignation aléatoire. On verra en fin de session les limites d'une telle approche et dans les suivantes des manières de construire des contrefactuels plus vraisemblables pour un sujet comme celui-ci.

On commence par charger les librairies requises (après les avoir installé si nécessaire)

```{r}
# On charge les librairies utiles pour cette analyse
library(tidyverse) # Facilite la manipulation de données
library(gt) # Aide à formater de jolis tableaux de rendu
library(broom) # Aide à formater les rendus de régressions
library(stargazer) # idem
library(sf) # Pour les données spatiales
library(lubridate) # Pour gérer des dates
library(readxl)
library(htmltools)

# Désactiver les notations scientifiques
options(scipen =999)

# On charge AP_Vahatra et on renomme la variable qui va nous servir pour la 
# jointure avec les données de Carvalho et al.
AP_Vahatra <- st_read("data/AP_Vahatra2.geojson") %>%
  rename(num_atlas = num_atlas_)
```

Pour commencer, on récupère le travail réalisé par Carvalho et al. [-@carvalho_methods_2020] qui complète les informations physiques de Goodman et al. [-@goodman_les_2018] avec des données relatives au couvert forestier en 1996, 2006 et 2016 et la diversité d'espèces présentes.

> Exercice : charger les données de Carvalho2018sup4.xlsx. Garder les données des lignes ayant TOTAL dans la colonne "Habitat type" Faire une jointure (left_join) pour inclure les données de Carvalho dans Vahatra

```{r}
sup2 <- read_xlsx("data/Carvalho2018sup2.xlsx", # Enplacement du fichier
                  skip = 8, # Premières lignes du tableau excel à ne pas lire
                  n_max = 101,  # on ne lit pas les dernières lignes (notes)
                  col_types = c("text", "text", "text", "text", 
        "numeric", "numeric", "numeric", "numeric", "numeric", "numeric", 
        "numeric", "numeric", "numeric", "numeric")) %>%
  mutate(`Site number` = as.numeric(`Site number`))
sup4 <- read_xlsx("data/Carvalho2018sup4.xlsx", skip = 6,
                  col_types = c("text", "numeric", "text", "numeric", "numeric", 
                                "numeric", "numeric", "numeric", "numeric", 
                                "numeric"))
sup4 <- sup4 %>%
  filter(`Habitat type` == "TOTAL") 

AP_Vahatra <- AP_Vahatra %>%
   left_join(sup2, by = c("num_atlas" = "Site number")) %>%
  left_join(sup4, by = c("num_atlas" = "Parcel"))
```

On commence par vérifier s'il existe des déséquilibres flagrants entre les aires qui ont été protégées avant 2015 et celles qui ont été protégées en 2015, en matière de surface totale ou de part couverte par des forêts en 1996.

```{r}
# On va renommer les variables du tableau final. 
rct_AP_Mada <- AP_Vahatra %>%
  st_drop_geometry() %>%
  rename(`Déforestation 1996-2006 (%)` = 
           `Forest loss (ha) between 1996-2006 (percent loss)`,
         `Déforestation 2006-2016 (%)` = 
           `Forest loss (ha) between 2006-2016 (percent loss)`,
         `Surface (ha)` = hectares_sum) %>%
  mutate(Groupe = ifelse(year(date_creation) < 2015, "Traitement", "Controle"),
         `Couvert forestier en 1996 (%)` = `Forest cover (ha) in 1996` / 
                                              `Surface (ha)` * 100,
         `Déforestation 1996-2016 (%)` = 
           (`Forest loss (ha) between 1996-2006 (absolute loss)` + 
             `Forest loss (ha) between 2006-2016 (absolute loss)`) /
           `Forest cover (ha) in 1996` * 100)

# On enregistre le jeu de données pour s'en resservir dans les prochaines 
# sessions
write_rds(rct_AP_Mada, "data/Vahatra_deforestation.rds")

# On fait une série de tests de comparaison de moyenne
t_tests <- rct_AP_Mada %>% 
  # On applique aux variables de déforestation, couvert en 96 et taille
  summarise(across(ends_with("(%)") | ends_with("(ha)"),# toutes finissent ainsi
                   ~ t.test(.[Groupe == "Controle"], # on applique un t.test
                            .[Groupe == "Traitement"])$p.value)) %>%
  mutate(Groupe = "t-test")

equilibre_avant <- rct_AP_Mada %>%
  group_by(Groupe) %>%
  summarise(`Nombre d'aires` = n(),
            `Sans forêt` = sum(is.na(`Couvert forestier en 1996 (%)`)), 
            `Surface (ha)` = mean(`Surface (ha)`),
            `Couvert forestier en 1996 (%)` = 
              mean(`Couvert forestier en 1996 (%)`, na.rm = TRUE)) %>%
  bind_rows(t_tests) %>% # On colle tous les t-tests 
  mutate(across(!Groupe, ~round(., 2))) %>%# arrondit tout sauf colonne "Groupe"
  select(-starts_with("Déforestation"))# On ne garde que les t-tests de baseline

# Ce qui suit est une série d'opération pour formater le rendu en tableau
equilibre_avant %>%
  t() %>% # On transpose lignes <=> colonnes
  as.data.frame() %>% # La transposition a altéré le format, on remet en tableau
  tibble::rownames_to_column() %>% # On met le nom des lignes en 1° colonne
  # "Truc pour renommer avec le contenu de la première ligne
  `colnames<-` (filter(., row_number() == 1)) %>% 
  filter(row_number() != 1)%>% # Enlève la 1° ligne qui est maintenant en entête
  gt() %>%
  tab_header(title = "Equilibre des variables avant intervention",
             subtitle = "(exercice : \"comme si\" c'était une RCT)") %>%
  tab_source_note("Source : Association Vahatra et Carvalho et al. 2018")
```

> Exercice : analysez le résultat de cette table

On a à première vue des déséquilibres limités "avant intervention". En moyenne, les deux groupes sont assez proches en termes de surface et de couvert forestier et le test de Student ne permet pas de rejeter l'hypothèse nulle concernant une différence de moyenne sur ces critères.

On va maintenant s'intéresser aux différences de déforestation observées "après intervention" dans le groupe de traitement.

```{r}

comparaison_apres <- rct_AP_Mada %>%
  group_by(Groupe) %>%
  summarise(across(starts_with("Déforestation"), ~ mean(., na.rm = TRUE))) %>%
  bind_rows(t_tests) %>% # On colle tous les t-tests 
  mutate(across(!Groupe, ~round(., 2))) %>%# arrondit tout sauf colonne "Groupe"
  select(Groupe, starts_with("Déforestation"))# On ne garde que les t-tests de baseline


# Même procédure que plus haut pour formater le rendu en tableau
comparaison_apres  %>%
  t() %>% # On transpose lignes <=> colonnes
  as.data.frame() %>% # La transposition a altéré le format, on remet en tableau
  tibble::rownames_to_column() %>% # On met le nom des lignes en 1° colonne
  # "Truc pour renommer avec le contenu de la première ligne
  `colnames<-` (filter(., row_number() == 1)) %>% 
  filter(row_number() != 1)%>% # Enlève la 1° ligne qui est maintenant en entête
  gt() %>%
  tab_header(title = "Moyennes après intervention",
             subtitle = "(exercice : \"comme si\" c'était une RCT)") %>%
  tab_source_note("Source : Association Vahatra et Carvalho et al. 2018")
```

> Exercice: commentez le résultat de cette table.

On peut également réaliser une régression simple, qu'on présente selon le format courant pour la littérature en économie grâce au package {stargazer} [@hlavac2022].

```{r}
#| output: false

# On exécute des régression pour chaque période séparément
def_96_06 <- lm(`Déforestation 1996-2006 (%)`  ~ Groupe, data = rct_AP_Mada)
def_06_16 <- lm(`Déforestation 2006-2016 (%)`  ~ Groupe, data = rct_AP_Mada)
def_96_16 <- lm(`Déforestation 1996-2016 (%)`  ~ Groupe, data = rct_AP_Mada)

# On consolide les résultats des régressions dans une table qu'on formate
# avec le package stargazer
rct_out1 <- stargazer(def_96_06, def_06_16, def_96_16, type = "html",
          title = "Impact de la conservation sur la perte de couvert forestier",
          notes = "Données : Association Vahatra et Carvalho et al. 2018")  

# Cette commande s'imprime aussi bien dans RStudio que dans le rendu quarto
browsable(HTML(paste(rct_out1, collapse = "")))
```

> Exercice : analyser le résultat de cette table. Qu'est-ce qu'elle suggère ?

On analyse ensuite la relation aux variables topologiques (altitude, indice de terrain accidenté) et de temps de trajet à la ville la plus proche en 2015. Le seuil retenu ici pour considérer une localité comme une ville est qu'elle ait au moins 5000 habitants.

```{r}
t_tests_autres <- rct_AP_Mada %>% 
  # On applique aux variables d'altitude, TRI et temps de trajet aux villes.
  summarise(across(minutes_mean:popcount_2000,# toutes finissent ainsi
                   ~ t.test(.[Groupe == "Controle"], # on applique un t.test
                            .[Groupe == "Traitement"])$p.value)) %>% 
  mutate(Groupe = "t-test")


equilibre_autres <- rct_AP_Mada %>%
  group_by(Groupe) %>%
  summarise(`Nombre d'aires` = n(),
            tri_mean = mean(tri_mean, na.rm = TRUE), 
            minutes_mean = mean(minutes_mean, na.rm = TRUE),
            elevation_mean = mean(elevation_mean, na.rm = TRUE),
            popcount_2000 = mean(popcount_2000, na.rm = TRUE)) %>%
  bind_rows(t_tests_autres) %>% # On colle tous les t-tests 
  mutate(across(!Groupe, ~round(., 2))) %>%# arrondit tout sauf colonne "Groupe"
  select(-starts_with("Déforestation"))# On ne garde que les t-tests de baseline

equilibre_autres %>%
  gt() %>%
  tab_header(title = "Equilibre entre les groupes en matière topologique",
             subtitle = "(exercice : \"comme si\" c'était une RCT)") %>%
  tab_source_note("Source : Nasa SRTM, Nelson et al.")

write_rds(rct_AP_Mada, file = "data/rct_AP_Mada.rds")
```

Le temps de trajet aux villes est significativement distinct entre les deux groupes. Attention, car cette variable pose un problème d'endogénéité, car le jeu de données utilisé pour cela date de 2015, alors que notre période d'étude démarre en 1996. Or, il est possible que la présence d'aires protégées ait eu une incidence sur la construction ou l'amélioration des tronçons routiers à proximité (discussion en séance sur ce point). Il semble important de tenir compte de l'accessibilité géographiques des aires protégées, mais une donnée antérieure à 2015 serait préférable.

On essaye de limiter ce biais en ajoutant le temps de trajet à une ville comme variables de contrôle à notre régression.

```{r}
#| output: false

def_96_16_controle <- lm(`Déforestation 1996-2016 (%)` ~ 
                           Groupe + minutes_mean, 
                         data = rct_AP_Mada)

rct_out2 <- stargazer(def_96_16, def_96_16_controle, type = "html")
browsable(HTML(paste(rct_out2, collapse = "")))

# Dans un bloc plus bas et non affiché, on a le code suivant
# cat(rct_out1)
```

> Exercice : interprétez le résultat
